refactor: SocialProvider, SocialUserInfo 패키징 이동
