+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/dto/response/AppleUserInfoResponse.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.account.auth.oauth.dto.response;
 
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
+import com.mobble.mobbleserver.application.account.command.SocialProvider;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
