+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/dto/response/KakaoUserInfoResponse.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.account.auth.oauth.dto.response;
 
-import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
+import com.mobble.mobbleserver.application.account.command.SocialProvider;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
