+++ src/main/java/com/mobble/mobbleserver/application/account/command/SocialUserInfo.java
@@ -1,0 +1,8 @@
+package com.mobble.mobbleserver.application.account.command;
+
+public record SocialUserInfo(
+        String email,
+        SocialProvider socialProvider,
+        String socialId
+) {
+}
