+++ src/main/java/com/mobble/mobbleserver/infrastructure/jwt/TokenProvider.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.infrastructure.jwt;
 
-import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
-import com.mobble.mobbleserver.account.auth.oauth.verifier.dto.SocialUserInfo;
+import com.mobble.mobbleserver.application.account.command.SocialProvider;
+import com.mobble.mobbleserver.application.account.command.SocialUserInfo;
 import com.mobble.mobbleserver.domain.clubMember.ClubMemberRole;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
