+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/account/dto/request/SocialLoginRequestDto.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.infrastructure.web.account.dto.request;
 
-import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
+import com.mobble.mobbleserver.application.account.command.SocialProvider;
 
 public record SocialLoginRequestDto(String accessToken, SocialProvider socialProvider) {
 }
