+++ src/test/java/com/mobble/mobbleserver/domain/member/MemberTest.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.member;
 
-import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
+import com.mobble.mobbleserver.application.account.command.SocialProvider;
 import com.mobble.mobbleserver.domain.ground.Ground;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
