+++ src/test/java/com/mobble/mobbleserver/infrastructure/persistence/member/MemberRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.infrastructure.persistence.member;
 
-import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
+import com.mobble.mobbleserver.application.account.command.SocialProvider;
 import com.mobble.mobbleserver.config.QueryDslConfig;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.support.fixture.member.MemberTestFixture;
