+++ src/test/java/com/mobble/mobbleserver/support/fixture/member/MemberTestFixture.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.support.fixture.member;
 
-import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
+import com.mobble.mobbleserver.application.account.command.SocialProvider;
 import com.mobble.mobbleserver.domain.member.Gender;
 import com.mobble.mobbleserver.domain.member.Member;
